lpm-levels: Use residency instead of power and energy overhead values
The driver looks through all the enabled modes and does energy
calculation runtime before it selects a low power mode to enter. With
residency values, the cpu/cluster can choose the low power mode as soon
as it comes across the low power mode which meets the residency
constraint.

Change-Id: I71235c212b26c3688a8ac5426942f2b88b681d47
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>

Conflicts:
	drivers/cpuidle/lpm-levels.c