pme: bcmdhd: Sanity check the HTC driver hang check
If there is a 10 second lag in xmit handling, the wifi device
is considered hung and "restarted" [*].

Unfortunately, when disabling wifi and then later re-enabling it,
it is possible (but I don't know exactly why) for this check to
hang test to declare a false positive.

To reproduce, I found that immediately after a reboot running a
speedtest on wifi, disabling wifi, running a speedtest on mobile
data, enabling wifi triggered this fairly frequently.

Prior to this commit running that scenario I would see the stuck
message in the logcat and following that wifi would be unrecoverable.
After this commit, I have seen the new potentially stuck message
and wifi continues to work.

[*]: Independent of this commit, the driver assumes that userspace
is handling a HANG message which we aren't.  If we were handling that
message correctly, this would probably just involve a slight wifi
lag and not the complete destruction of wifi connectivity.

Change-Id: I989c567bfe44fd5ff5942049205b2ac318c015db